ROL: Separate the code into two classes - backend and frontend
- Backend is the AdLib Visual Composer synth class,
  which will be reused for all derived music formats.

- Frontend is the file format player only,
  talking to backend via its public interface.

Making a separate backend class will help with code deduplication,
as it could (and should) be used instead of adlib.c/adlib.h driver.
